requirements-dev.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==5.0.0
black==24.8.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
Faker==37.12.0
fastapi==0.112.2
filelock==3.20.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==5.13.2
loguru==0.7.3
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre-commit==3.8.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pytest==8.2.2
pytest-asyncio==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
ruff==0.6.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.30.5
virtualenv==20.35.4
win32_setctime==1.2.0
